% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chernoff_bound.R
\name{chernoff_bound}
\alias{chernoff_bound}
\title{Calculates the chernoff bound simulations.}
\usage{
chernoff_bound(rpart.tree, n_experimentations, epsilon)
}
\arguments{
\item{rpart.tree}{rpart.tree. A Decision tree generated by rpart package.}

\item{n_experimentations}{int. The # of experimentations and samples to run.}

\item{epsilon}{float. The epsilon to be used in the delta calculation.}
}
\value{
No return value, the function plots the chernoff bound.
}
\description{
Calculates different values of shattering coefficient and delta,
varying # of samples to study the chernoff bound of SLT.
}
