% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_shattering.R
\name{compute_shattering}
\alias{compute_shattering}
\title{Calculates the shattering coefficient for a decision tree.}
\usage{
compute_shattering(rpart.tree, n_samples)
}
\arguments{
\item{rpart.tree}{rpart.tree. A Decision tree generated by rpart package.}

\item{n_samples}{int. The # of samples to consider in dataset.}
}
\value{
The calculus of the shattering coefficient.
}
\description{
Using a recursion function as auxiliar function,
calculates the shattering coefficient for a rpart decision tree.
}
