% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_interval.R
\name{confidence_interval}
\alias{confidence_interval}
\title{Calculates the confidence interval for the dataset.}
\usage{
confidence_interval(rpart.tree, epsilon, max_confide)
}
\arguments{
\item{epsilon}{float. The epsilon to be used in the delta calculation.}

\item{max_confidence}{int. The maximum confidence to achieve in 0.01 steps length.}

\item{rpart_tree}{rpart.tree. A Decision tree generated by rpart package.}
}
\description{
Calculates # of samples needed to ensure learning with,
a certain confidence interval simulated.
}
\examples{
confidence_interval(rpart.tree)

}
