% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{g}
\alias{g}
\title{Apply G(n) function.}
\usage{
g(n, n_samples)
}
\arguments{
\item{n}{float. The number to compute G}

\item{n_samples}{integer. The # of samples in dataset.}
}
\value{
The calculus of G function for the inputs.
}
\description{
Apply G(n) function on a given number (n) for a
given # of samples (n_samples).
}
\examples{
g(2, 10)
g(3, 15)

}
