% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurse.R
\name{recurse}
\alias{recurse}
\title{Calculates the shattering coefficient for a decision tree.}
\usage{
recurse(nodes, left, right, node, parent, samples, node_type, shattering, n_samples)
}
\arguments{
\item{nodes}{list. A list containing the tree nodes.}

\item{left}{int. The left child of the parent node.}

\item{right}{int. The right child of the parent node.}

\item{node}{int. The current node being analyzed}

\item{parent}{int. The parent node of node.}

\item{samples}{list. The number of dataset samples for each node of the tree.}

\item{node_type}{string. The type of the node, e.g., leaf if it's a leaf node.}

\item{shattering}{float. The shattering coefficient that are begin computed.}

\item{n_samples}{int. The # of samples to consider in dataset.}
}
\value{
The calculus of the shattering coefficient.
}
\description{
Using a recursion, calculates the shattering coefficient
for a rpart decision tree.
}
