% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_n_samples.R
\name{search_n_samples}
\alias{search_n_samples}
\title{Executes a binary search to find the best # of samples.}
\usage{
search_n_samples(rpart.tree, lower, upper, epsilon)
}
\arguments{
\item{lower}{int. The # of samples to be used in lower_delta calculation.}

\item{upper}{int. The # of samples to be used in upper_delta calculation.}

\item{epsilon}{float. The epsilon to be used in the delta calculation.}

\item{rpart_tree}{rpart.tree. A Decision tree generated by rpart package.}
}
\value{
the number of samples needed to ensure learning.
}
\description{
Executes a binary search to find the best # of samples.
}
\examples{
search_n_samples(rpart.tree, 16, 64, 0.05)

}
