% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shattering_simulations.R
\name{shattering_simulations}
\alias{shattering_simulations}
\title{Calculates the shattering coefficient simulations.}
\usage{
shattering_simulations(rpart.tree)
}
\arguments{
\item{n_experimentations}{int. The # of experimentations and samples to run.}

\item{rpart_tree}{rpart.tree. A Decision tree generated by rpart package.}
}
\description{
Calculates different values of shattering coefficient,
varying # of samples to study the # of samples needed to ensure learning
and plot this values.
}
\examples{
shattering_simulations(rpart.tree)

}
