% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_delta.R
\name{compute_delta}
\alias{compute_delta}
\title{Calculates the delta for a given # of samples and value of epsilon.}
\usage{
compute_delta(rpart.tree, samples, epsilon)
}
\arguments{
\item{rpart.tree}{rpart.tree. A Decision tree generated by rpart package.}

\item{samples}{int. The # of samples to be used in delta calculation.}

\item{epsilon}{float. The epsilon to be used in the delta calculation.}
}
\value{
computed delta for # of samples and epsilon given as input.
}
\description{
Calculates the delta for a given # of samples and value of epsilon.
}
