% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_interval.R
\name{confidence_interval}
\alias{confidence_interval}
\title{Calculates the confidence interval for the dataset.}
\usage{
confidence_interval(rpart.tree, epsilon, max_confidence)
}
\arguments{
\item{rpart.tree}{rpart.tree. A Decision tree generated by rpart package.}

\item{epsilon}{float. The epsilon to be used in the delta calculation.}

\item{max_confidence}{int. The maximum confidence to achieve in 0.01 steps length.}
}
\value{
No return value, the function plot the confidence interval values.
}
\description{
Calculates # of samples needed to ensure learning with,
a certain confidence interval simulated.
}
