% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_delta_n_samples.R
\name{search_delta_n_samples}
\alias{search_delta_n_samples}
\title{Search the # of samples to ensure learning given an epsilon.}
\usage{
search_delta_n_samples(rpart.tree, n_samples, delta, epsilon)
}
\arguments{
\item{rpart.tree}{rpart.tree. A Decision tree generated by rpart package.}

\item{n_samples}{int. The # of samples to be used as initial value to search the minimum necessary to ensure learning.}

\item{delta}{float. The initial value of delta (need to be greater than epsilon)}

\item{epsilon}{float. The epsilon to be used in the delta calculation.}
}
\value{
the delta and n_samples values.
}
\description{
Search the # of samples to ensure learning given an epsilon.
}
