% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shattering_simulations.R
\name{shattering_simulations}
\alias{shattering_simulations}
\title{Calculates the shattering coefficient simulations.}
\usage{
shattering_simulations(rpart.tree, n_experimentations=200000)
}
\arguments{
\item{rpart.tree}{rpart.tree. A Decision tree generated by rpart package.}

\item{n_experimentations}{int. The # of experimentations and samples to run.}
}
\value{
No return value, the function plot shattering simulations.
}
\description{
Calculates different values of shattering coefficient,
varying # of samples to study the # of samples needed to ensure learning
and plot this values.
}
